# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Copernicus",
    "description": "ASR + NLP text correction service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tasks/{task_id}/results": {
      "get": {
        "tags": ["tasks"],
        "summary": "Get Task Results",
        "description": "返回任务的所有持久化结果。",
        "operationId": "get_task_results_api_v1_tasks__task_id__results_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResultsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/evaluate/text/async": {
      "post": {
        "tags": ["evaluation"],
        "summary": "Submit Text Evaluation Task",
        "description": "提交异步文本评估任务，通过 GET /tasks/{task_id} 轮询进度。",
        "operationId": "submit_text_evaluation_task_api_v1_evaluate_text_async_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_text_evaluation_task_api_v1_evaluate_text_async_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_submit_text_evaluation_task_api_v1_evaluate_text_async_post": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "parent_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Task Id"
          }
        },
        "type": "object",
        "required": ["text"],
        "title": "Body_submit_text_evaluation_task_api_v1_evaluate_text_async_post"
      },
...
      "ComplianceReport": {
        "properties": {
          "total_rules": {
            "type": "integer",
            "title": "Total Rules"
          },
          "total_segments_checked": {
            "type": "integer",
            "title": "Total Segments Checked"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/Violation"
            },
            "type": "array",
            "title": "Violations"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "compliance_score": {
            "type": "number",
            "title": "Compliance Score",
            "default": 100
          },
          "source_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Source Counts"
          }
        },
        "type": "object",
        "required": ["total_rules", "total_segments_checked"],
        "title": "ComplianceReport",
        "description": "Full compliance audit report."
      },
      "ComplianceResponse": {
        "properties": {
          "rules": {
            "items": {
              "$ref": "#/components/schemas/ComplianceRule"
            },
            "type": "array",
            "title": "Rules"
          },
          "report": {
            "$ref": "#/components/schemas/ComplianceReport"
          },
          "processing_time_ms": {
            "type": "number",
            "title": "Processing Time Ms"
          }
        },
        "type": "object",
        "required": ["rules", "report", "processing_time_ms"],
        "title": "ComplianceResponse",
        "description": "API response for compliance audit."
      },
      "ComplianceRule": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": ["id", "content"],
        "title": "ComplianceRule",
        "description": "Single audit rule parsed from CSV/XLSX."
      },
      "EvaluationAnalysis": {
        "properties": {
          "main_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Main Points"
          },
          "key_data": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Data"
          },
          "sentiment": {
            "type": "string",
            "title": "Sentiment",
            "default": ""
          }
        },
        "type": "object",
        "title": "EvaluationAnalysis"
      },
      "EvaluationMeta": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": ""
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          }
        },
        "type": "object",
        "title": "EvaluationMeta"
      },
...
      "EvaluationResult": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/EvaluationMeta"
          },
          "scores": {
            "$ref": "#/components/schemas/EvaluationScores"
          },
          "analysis": {
            "$ref": "#/components/schemas/EvaluationAnalysis"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          }
        },
        "type": "object",
        "title": "EvaluationResult"
      },
      "EvaluationScores": {
        "properties": {
          "logic": {
            "type": "integer",
            "title": "Logic",
            "default": 0
          },
          "info_density": {
            "type": "integer",
            "title": "Info Density",
            "default": 0
          },
          "expression": {
            "type": "integer",
            "title": "Expression",
            "default": 0
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          }
        },
        "type": "object",
        "title": "EvaluationScores"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskResultsResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "transcript": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranscriptResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "compliance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ComplianceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "has_audio": {
            "type": "boolean",
            "title": "Has Audio",
            "default": false
          },
          "has_video": {
            "type": "boolean",
            "title": "Has Video",
            "default": false
          },
          "keyframe_count": {
            "type": "integer",
            "title": "Keyframe Count",
            "default": 0
          },
          "ocr_text_count": {
            "type": "integer",
            "title": "Ocr Text Count",
            "default": 0
          },
          "visual_event_count": {
            "type": "integer",
            "title": "Visual Event Count",
            "default": 0
          }
        },
        "type": "object",
        "required": ["task_id"],
        "title": "TaskResultsResponse",
        "description": "Persisted results for a task (used for restoring state on page refresh)."
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing_asr",
          "extracting_frames",
          "scanning_visual",
          "correcting",
          "evaluating",
          "auditing",
          "completed",
          "failed"
        ],
        "title": "TaskStatus"
      },
...
      "TaskSubmitResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "existing": {
            "type": "boolean",
            "title": "Existing",
            "default": false
          }
        },
        "type": "object",
        "required": ["task_id", "status"],
        "title": "TaskSubmitResponse"
      },
      "TranscriptEntrySchema": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "timestamp_ms": {
            "type": "integer",
            "title": "Timestamp Ms"
          },
          "end_ms": {
            "type": "integer",
            "title": "End Ms",
            "default": 0
          },
          "speaker": {
            "type": "string",
            "title": "Speaker"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "text_corrected": {
            "type": "string",
            "title": "Text Corrected"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "timestamp_ms",
          "speaker",
          "text",
          "text_corrected"
        ],
        "title": "TranscriptEntrySchema"
      },
      "TranscriptResponse": {
        "properties": {
          "transcript": {
            "items": {
              "$ref": "#/components/schemas/TranscriptEntrySchema"
            },
            "type": "array",
            "title": "Transcript"
          },
          "processing_time_ms": {
            "type": "number",
            "title": "Processing Time Ms"
          }
        },
        "type": "object",
        "required": ["transcript", "processing_time_ms"],
        "title": "TranscriptResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "Violation": {
        "properties": {
          "rule_id": {
            "type": "integer",
            "title": "Rule Id"
          },
          "rule_content": {
            "type": "string",
            "title": "Rule Content"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "severity": {
            "type": "string",
            "enum": ["high", "medium", "low"],
            "title": "Severity",
            "default": "low"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "rejected"],
            "title": "Status",
            "default": "pending"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "default": ""
          },
          "timestamp_ms": {
            "type": "integer",
            "title": "Timestamp Ms",
            "default": 0
          },
          "end_ms": {
            "type": "integer",
            "title": "End Ms",
            "default": 0
          },
          "speaker": {
            "type": "string",
            "title": "Speaker",
            "default": ""
          },
          "original_text": {
            "type": "string",
            "title": "Original Text",
            "default": ""
          },
          "source": {
            "type": "string",
            "enum": ["transcript", "ocr", "vision"],
            "title": "Source",
            "default": "transcript"
          },
          "evidence_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Url"
          },
          "evidence_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Text"
          },
          "rule_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Ref"
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          }
        },
        "type": "object",
        "required": ["rule_id", "rule_content", "reason", "confidence"],
        "title": "Violation",
        "description": "Single violation detected by LLM."
      },
...
    }
  }
}